+++ src/Frontend/Dfe.Complete/Pages/Projects/Team/ProjectsByUser/ProjectsForUser.cshtml.cs
@@ -1,0 +1,28 @@
+﻿using Dfe.Complete.Application.Users.Models;
+using Dfe.Complete.Application.Users.Queries.GetUser;
+using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
+using Dfe.Complete.Models;
+using Dfe.Complete.Pages.Pagination;
+using MediatR;
+
+namespace Dfe.Complete.Pages.Projects.Team.ProjectsForUser;
+
+public class ProjectsForUser(ISender sender) : YourTeamProjectsModel(ByUserNavigation)
+{
+    public UserWithProjectsDto UserResult { get; set; } = default!;
+
+    public async Task OnGet(string id)
+    {
+        ViewData[TabNavigationModel.ViewDataKey] = YourTeamProjectsTabNavigationModel;
+
+        var getUserQuery = new GetUserWithProjectsQuery(new UserId(new Guid(id)), ProjectState.Active)
+            { Page = PageNumber - 1, Count = PageSize };
+
+        var response = await sender.Send(getUserQuery);
+        UserResult = response.Value ?? default!;
+        
+        Pagination = new PaginationModel($"/projects/team/users/{id}", PageNumber, response.ItemCount, PageSize);
+
+    }
+}